reinstate TestimonyTab
